move all helper  functions from views module to utils module, update main.py, update utils and views tests
